Fix build dependency on deprecated msgpack unpacker api call.
As this project has been idle for awhile it looks like a deprecated
method on msgpack::unpacker was causing the build to break for the
latest and greatest msgpack-c library. More specifically the method
msgpack::unpacker::rewind_buffer no longer exists. In reading the
current msgpack-c code it looks like the correct fix here is to
simply remove the call to rewind_buffer (refer to unpack.h).