Require BOOST_ASIO_SEPARATE_COMPILATION for static libraries.
This is consistent with requiring BOOST_ASIO_DYN_LINK for shared
libraries, and makes more sense for being a library.

Requires the library user to include <boost/asio/impl/src.hpp>
in a separate compilation unit, ideally in the top-level project,
which is then picked up by the (static or shared) library.