Remove timers and signals from the event loop.
Using a single boost timer in the event loop only works if you are
only ever adding a single timer. If you try to add 2 timers then the
first registered handler is cancelled and the new timer handler is
added. Of course the original timer handler re-registers itself upon
cancellation which causes the second timer handler to be cancelled and
so on. This simply results in a cpu core being consumed by a cylce of
timer handler cancellation and re-registeration. Therefore I have pulled
the timer out of the event loop and now we just create boost deadline
timers in the appropriate context in which they are needed. This actually
simplifies the step timeouts that the timers are actually used for. While
I was at it I also purged the signal part of the event loop as well as it
wasn't used and would innevitably suffer the same fate as dealine timers.